Clarify that the --interactive step can be avoided
The section that describes running an initial `dotnet restore --interactive` command to acquire a token is worded in a way that implies that this step is mandatory. In fact it isn't because it is also possible to use the `VSS_NUGET_EXTERNAL_FEED_ENDPOINTS` environment variable to supply you own access token if you already have one in hand.

It was really easy to miss that because the first section makes it sound like only interactive operation is supported, and even if that doesn't put you off and you continue to look for a way, there's nothing in either the "Environment Variables" title or the `VSS_NUGET_EXTERNAL_FEED_ENDPOINTS` variable name to give you a clear clue that this provides a way to bypass interactive operation.

By adding an extra paragraph directly after the description of the interactive process, it is easier for readers to discover that an alternative exists.