updated install from cli example
Received an "Access Denied" error when attempting to use the install from cli example.  Also tried it as an administrator.  Found a Microsoft article explaining that iex shouldn't be used: https://devblogs.microsoft.com/powershell/invoke-expression-considered-harmful/

Changed it use to wget to fetch and then run the script.  Downloaded and ran the installer without issue.  Also included powershell in the example because not every dev would know to prefix that without being told to do so.