Fix credential provider release build and resolve build warnings (#354)
The release build for ingestion to Visual Studio was broken by service connection remediation so needed to be recreated. Some recent changes in the build templates are now no longer compatible with the build machines running release builds. Also fixing some miscellaneous build and release warnings that have snuck in.