Add devcontainer.json and .vscode defaults (#398)
Add default devcontainer.json to enable the repository to build in
Codespaces. This builds from the dotnet latest images and adds .NET 6.0.
The repository is able to build, run, and test if using .NET 6.0
(probably .NET Core 3.1 as well but untested). .NET Framework requires
mono to run but didn't see any convenient features for that.

Also adding the artifacts-helper to set up NuGet authentication so
collaborators on the Azure DevOps feed are able to pull new packages
into the feed.