Add .NET 8, FX 4.8.1 support (#520)
With the end of support for multiple .NET releases used by the artifacts
credential provider, it's time to migrate to supported .NET versions.
This includes a bump of .NET 6 -> .NET 8 and .NET Framework 4.6.1 to
.NET Framework 4.8.1.

For more info on end of support dates, see
https://github.com/dotnet/core/blob/main/releases.md

These new bits will not be set as the default version being installed
until the next major release as documented in the README. I'll also post
an announcement about the upcoming deprecation.

---------

Co-authored-by: Coby Allred <coallred@microsoft.com>