Update installscript.sh to clean content before extractions (#611)
This pull request updates the installation script for the Microsoft
NuGet credential provider to ensure a clean installation. The main
change is the addition of a step that removes any existing credential
provider before installing a new one.

Credential provider installation improvements:

* Added logic to `helpers/installcredprovider.sh` to remove any existing
`CredentialProvider.Microsoft` directory from
`${NUGET_PLUGIN_DIR}/netcore` before proceeding with the installation,
ensuring that outdated or conflicting files do not interfere with the
new installation.

This was brought to my attention by issue #610